Changed fields access modifiers to private if only used within the Kiosk class.
Removed getters as they were redudant to the program.

Adjusted the areas that called getters to simply call the associated field instead.

Changed the initial prompt when the user picks the card payment method to simply ask them to enter their card number. The previous branch made a confusing statement by telling the customer that the card number is invalid before any input was made.

Made the error prompt for invalid input of the card number look neater when presented on the console.

Made the process of cash payments look neater when presented to the console. The console will now give better context to the user of much is still owed on the purchase at all steps of the payment process.

Changed the MoneyRequest method to a static method

Removed the refund transaction method as this is not required to simulate for this project.

Moved the post transaction kiosk report into a method within the Kiosk class.

Moved the logic for running an external program and arguments passed to that program into a method within the Kiosk class.

Renamed the LogTransaction project's .cs file to match the project name.

Made minor spacing adjustments to the LogTransaction script.