Fixed an error where the user could not reinput a card number when their initial card transaction attempt declined. An error was caused by not resetting the digit count to zero when the do while loop was repeated.
Removed unused fields within the Kiosk class.

Changed the formatting of the _cardVendor variable when the card vendor is American Express to be "AmericanExpress". This prevents an exception from being thrown when the _cardVendor variable is passed as an argument when the card vendor name is more than one word.

Added logic to LogTransaction.cs to reformat the cardVendor string to have proper spacing when the card vendor is "American Express".

Minor changes to developer comments and console output for better context.