All fields of class Kiosk are now set to read-only.
Moved the string variables that contain the .log files paths within the LogTransaction project to the top with the rest of the variables in Main().

Cleaned up the spacing of all lines of the project's scripts.

Added comments to string arrays "text" and "attrib" to give context to their differences.