Set Kiosk inventory back to 5 of each bill and coin. This was an unintended change made in the previous branch.
Added the option to change payment method from card to cash if the user's card declined while the user wanted cash back. Note: adding a cash payment method as an alternate option for the previously described scenario made the output to the user one-to-one with the scenario in which their card declines and does not want cash back. Knowing this, there is no longer a reason to keep track of whether or not the user wants cash back using a boolean variable. An else-if statement that utilized this logic has also been removed to reduce arbitrary code.

Added some code that changes the value of booleans 'pass' and 'declined' to false and true respectively to force a card to decline for testing purposes. This should be commented out when pushing new branches to the repository.